Handle SwiftPackageList changes in tests
